docker_compose/fast-api-app/schema/schema.py
from pydantic import BaseModel


class UserBase(BaseModel):
    name: str
    email: str
    password: str


class UserCreate(UserBase):
    pass


class UserUpdate(UserBase):
    pass


class User(UserBase):
    id: int

    class Config:
        from_attributes = True
