docker_compose/fast-api-app/pyproject.toml
[project]
name = "fast-api-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.2",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
